+++ src/components/CaptureMatching/Grower.js
@@ -2,19 +2,24 @@ import React from 'react';
 
 import { Typography, Box } from '@material-ui/core';
 
-function Grower(props) {
+function Grower({ planter_photo_url, planter_username, status }) {
   return (
     <Box style={{ display: 'flex' }}>
       <img
-        src={props.userPhotoUrl}
+        src={planter_photo_url}
         alt=""
-        style={{ borderRadius: '50%', width: '130px', height: '130px' }}
+        style={{
+          borderRadius: '50%',
+          width: '130px',
+          height: '130px',
+          paddingTop: '20px',
+        }}
       />
       <Box style={{ margin: 'auto', paddingLeft: '10px' }}>
         <Typography variant="body1" style={{ display: 'block' }}>
-          Grower Name
+          {planter_username}
         </Typography>
-        <Typography variant="body1">grower status </Typography>
+        <Typography variant="body1">{status}</Typography>
       </Box>
     </Box>
   );