+++ src/context/AppContext.js
@@ -5,11 +5,12 @@ import axios from 'axios';
 import VerifyView from '../views/VerifyView';
 import GrowersView from '../views/GrowersView';
 import CapturesView from '../views/CapturesView';
+import EarningsView from '../views/EarningsView/EarningsView';
 import Account from '../components/Account';
 import Home from '../components/Home/Home';
 import Users from '../components/Users';
 import SpeciesView from '../views/SpeciesView';
-import CaptureMatchingFrame from '../components/CaptureMatching/CaptureMatchingFrame';
+import CaptureMatchingView from '../components/CaptureMatching/CaptureMatchingView';
 import Unauthorized from '../components/Unauthorized';
 
 import IconSettings from '@material-ui/icons/Settings';
@@ -23,6 +24,7 @@ import IconPermIdentity from '@material-ui/icons/PermIdentity';
 import CategoryIcon from '@material-ui/icons/Category';
 import HomeIcon from '@material-ui/icons/Home';
 import CompareIcon from '@material-ui/icons/Compare';
+import CreditCardIcon from '@material-ui/icons/CreditCard';
 import { session, hasPermission, POLICIES } from '../models/auth';
 import api from '../api/treeTrackerApi';
 
@@ -68,7 +70,7 @@ function getRoutes(user) {
     {
       name: 'Capture Matching',
       linkTo: '/capture-matching',
-      component: CaptureMatchingFrame,
+      component: CaptureMatchingView,
       icon: CompareIcon,
       disabled:
         process.env.REACT_APP_ENABLE_CAPTURE_MATCHING !== 'true' ||
@@ -76,6 +78,13 @@ function getRoutes(user) {
           POLICIES.SUPER_PERMISSION,
           POLICIES.APPROVE_TREE,
         ]),
+    },
+    {
+      name: 'Earnings',
+      linkTo: '/earnings',
+      component: EarningsView,
+      icon: CreditCardIcon,
+      disabled: process.env.REACT_APP_ENABLE_EARNINGS !== 'true',
     },
     {
       name: 'Growers',